change location of administration page files
