+++ src/PrestaShopBundle/Form/Admin/Configure/AdvancedParameters/Administration/NotificationsType.php
@@ -23,14 +23,14 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
-namespace PrestaShopBundle\Form\Admin\AdvancedParameters\Administration;
+namespace PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Administration;
 
+use PrestaShopBundle\Form\Admin\Type\SwitchType;
 use PrestaShopBundle\Form\Admin\Type\TranslatorAwareType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use PrestaShopBundle\Form\Admin\Type\TextWithUnitType;
 
-class UploadQuotaType extends TranslatorAwareType
+class NotificationsType extends TranslatorAwareType
 {
     /**
      * {@inheritdoc}
@@ -38,17 +38,14 @@ class UploadQuotaType extends TranslatorAwareType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('max_size_attached_files', TextWithUnitType::class, array(
+            ->add('show_notifs_new_orders', SwitchType::class, array(
                 'required' => true,
-                'unit' => $this->trans('megabytes', 'Admin.Advparameters.Feature'),
             ))
-            ->add('max_size_downloadable_product', TextWithUnitType::class, array(
+            ->add('show_notifs_new_customers', SwitchType::class, array(
                 'required' => true,
-                'unit' => $this->trans('megabytes', 'Admin.Advparameters.Feature'),
             ))
-            ->add('max_size_product_image', TextWithUnitType::class, array(
+            ->add('show_notifs_new_messages', SwitchType::class, array(
                 'required' => true,
-                'unit' => $this->trans('megabytes', 'Admin.Advparameters.Feature'),
             ))
         ;
     }
@@ -68,6 +65,6 @@ class UploadQuotaType extends TranslatorAwareType
      */
     public function getBlockPrefix()
     {
-        return 'administration_upload_quota_block';
+        return 'administration_notification_block';
     }
 }
