+++ src/PrestaShopBundle/Resources/config/services/form/form_data_provider.yml
@@ -33,7 +33,7 @@ services:
             - '@prestashop.adapter.maintenance.configuration'
 
     prestashop.adapter.administration.form_provider:
-        class: 'PrestaShopBundle\Form\Admin\AdvancedParameters\Administration\FormDataProvider'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\Administration\FormDataProvider'
         arguments:
             - '@prestashop.adapter.general.configuration'
             - '@prestashop.adapter.upload_quota.configuration'
