+++ src/PrestaShopBundle/Resources/views/Admin/Configure/AdvancedParameters/administration.html.twig
